Delete old GoModGraph and GoList tactics.
