+++ spectrometer.cabal
@@ -369,10 +369,8 @@ library
     Strategy.Fpm
     Strategy.Glide
     Strategy.Go.GlideLock
-    Strategy.Go.GoList
     Strategy.Go.GoListPackages
     Strategy.Go.Gomod
-    Strategy.Go.GoModGraph
     Strategy.Go.GopkgLock
     Strategy.Go.GopkgToml
     Strategy.Go.Gostd
@@ -566,8 +564,6 @@ test-suite unit-tests
     Fossa.API.TypesSpec
     Go.GlideLockSpec
     Go.GoListPackagesSpec
-    Go.GoListSpec
-    Go.GoModGraphSpec
     Go.GomodSpec
     Go.GopkgLockSpec
     Go.GopkgTomlSpec
