+++ test/Go/GoListPackagesSpec.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE QuasiQuotes #-}
+
 module Go.GoListPackagesSpec (
   spec,
 ) where
@@ -18,9 +20,9 @@ import Graphing qualified (Graphing, direct, edge)
 import Path (Abs, Dir, Path)
 import Path.IO (getCurrentDir)
 import ResultUtil (assertOnSuccess)
-import Strategy.Go.GoListPackages (GoModule (..), GoPackage (..), ImportPath (..), ModulePath (ModulePath), buildGraph)
-import Strategy.Go.GoModGraph (toGoModVersion)
+import Strategy.Go.GoListPackages (GoModule (..), GoPackage (..), ImportPath (..), ModulePath (ModulePath), buildGraph, toGoModVersion)
 import Test.Hspec (Spec, describe, it, runIO)
+import Text.RawString.QQ (r)
 
 -- In this set of packages there are two main modules.
 -- In the resulting graph expect each main module to be absent, with it's dependencies
