--- test/Go/testdata/go.mod.graph.stdout
@@ -1,2 +1,0 @@
-main dep@v0.5.0
-dep@v0.5.0 depB@v0.6.0
